% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGALTri-package.R
\docType{package}
\name{CGALTri-package}
\alias{CGALTri}
\alias{CGALTri-package}
\title{CGALTri: Triangulation with CGAL}
\description{
Perform triangulation in R using the CGAL library.
}
\details{
Experimental package that computes triangulation using CGAL.
}
\references{
CGAL, \emph{Computational Geometry Algorithms Library}, \url{https://www.cgal.org}.
}
\author{
\strong{Maintainer}: Peng Xu \email{px2132@columbia.edu}

}
\keyword{internal}
