% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_2D.R
\name{Plot_Cropped_Voronoi_2D}
\alias{Plot_Cropped_Voronoi_2D}
\title{Plot a Voronoi Diagram with Edges Restricted to a Rectangle}
\usage{
Plot_Cropped_Voronoi_2D(data, bCoord = NULL)
}
\arguments{
\item{data}{Input coordinate matrix, of size \eqn{n} by \eqn{2}.}

\item{bCoord}{Coordinates of the box boundary.}
}
\description{
Computes the Delaunay triangulation of a set of points and prints the Voronoi edges restricted to a given rectangle.
}
\keyword{Voronoi}
